ai_content_detector/requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
tokenizers>=0.13.0
datasets>=2.14.0

# Image Processing
Pillow>=9.5.0
opencv-python>=4.7.0
numpy>=1.24.0
scipy>=1.10.0

# API and Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6

# UI
streamlit>=1.25.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utils and Data
pandas>=2.0.0
requests>=2.31.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Evaluation and Metrics
scikit-learn>=1.3.0
wandb>=0.15.0

# Model Export and Optimization
onnx>=1.14.0
onnxruntime>=1.15.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0